Update bill for friends.java

+++ bill for friends.java
@@ -4,8 +4,8 @@ public class q22{
     public static void main(String[] args) {
          Scanner sc = new Scanner(System.in);  
          System.out.println("Enter bill amount and no fo friends: "); 
-         double bill = sc.nextInt();
-         double nof = sc.nextInt();
+         double bill = sc.nextDouble();
+         double nof = sc.nextDouble();
          double split = bill/nof;
          System.out.println("Bill  fro Every person: RS." + split);
     }
